Created component folders, renamed styles -> styled; cleaned up imports
