+++ components/product/Product.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import { Image, View } from 'react-native';
-
-import { ProductBody } from '../styles/products';
-import { Button } from '../styles/shared';
+import { ProductBody } from '../../styled/product';
+import { Button } from '../../styled/shared';
 
 /**
  * @prop
@@ -20,7 +19,7 @@ function Product({ product, store, navigation }) {
       }>
       <View style={{ alignItems: 'center' }}>
         <Image
-          source={require('../assets/images/robot-dev.png')}
+          source={require('../../assets/images/robot-dev.png')}
           style={{ width: 50, height: 50, borderRadius: 80 / 2 }}
         />
         <ProductBody>{product.name}</ProductBody>
