+++ components/product/StoreProducts.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import { View } from 'react-native';
 import { ScrollView } from 'react-native-gesture-handler';
-import { Button, Title } from '../styles/shared';
-import { TextButton } from './BaseComponents';
+import { Button, Title } from '../../styled/shared';
+import { TextButton } from '../BaseComponents';
+import StoreCard from '../store/StoreCard';
 import Product from './Product';
-import StoreCard from './StoreCard';
 
 function filterFruit(product) {
   if (product) {
