+++ components/rewards/TabBarIcon.js
@@ -1,7 +1,6 @@
-import React from 'react';
 import { Ionicons } from '@expo/vector-icons';
-
-import Colors from '../constants/Colors';
+import React from 'react';
+import Colors from '../../constants/Colors';
 
 export default function TabBarIcon(props) {
   return (
