+++ lib/authUtils.js
@@ -104,7 +104,7 @@ export const updateCustomerPushTokens = function async(
 
     // In a rare case (or in development), may not have registered a pushToken initially
     if (tokenNames) {
-      for (i = 0; i < tokenNames.length; i++) {
+      for (i = 0; i < tokenNames.length; i += 1) {
         if (tokenNames[i] === currentToken) {
           isNew = false;
           break;
