+++ screens/auth/SignUpScreen.js
@@ -11,13 +11,12 @@ import {
   View
 } from 'react-native';
 import validatejs from 'validate.js';
-
-import { Input, SignUpContainer } from '../../styles/auth';
 import {
   checkForDuplicateCustomer,
   createCustomer,
   createPushToken
 } from '../../lib/authUtils';
+import { Input, SignUpContainer } from '../../styled/auth';
 
 // I abstracted portions of the validation flow into these files
 // but there's a weird bug "https://github.com/facebook/react-native/issues/4968"
