+++ screens/map/ProductDetailsScreen.js
@@ -1,9 +1,8 @@
 import React from 'react';
 import { Image, View } from 'react-native';
-
-import ProductInfo from '../../components/ProductInfo';
-import { ImageContainer } from '../../styles/products';
-import { Title } from '../../styles/shared';
+import ProductInfo from '../../components/product/ProductInfo';
+import { ImageContainer } from '../../styled/product';
+import { Title } from '../../styled/shared';
 
 class ProductDetailsScreen extends React.Component {
   constructor(props) {
