+++ screens/map/StoreListScreen.js
@@ -1,10 +1,9 @@
 import React from 'react';
-import { ScrollView } from 'react-native-gesture-handler';
 import { SearchBar } from 'react-native-elements'; // @tommypoa: Create styled-component for this
-
-import StoreCard from '../../components/StoreCard';
-import { Title } from '../../styles/shared';
-import { StoreModal } from '../../styles/stores';
+import { ScrollView } from 'react-native-gesture-handler';
+import StoreCard from '../../components/store/StoreCard';
+import { Title } from '../../styled/shared';
+import { StoreModal } from '../../styled/store';
 
 class StoreListScreen extends React.Component {
   constructor(props) {
