+++ styled/rewards.js
